This contains all conflict-free changes from upstream and our tip.
All changes are computed against lastUpstreamSha:
  - [replayio/chromium](https://github.com/replayio/chromium/commit/27d3765d341b09369006d030f83f582a29eb57ae)
  - [chromium/chromium](https://github.com/chromium/chromium/commit/27d3765d341b09369006d030f83f582a29eb57ae)
ourTip: aa1324dd45b27a464154f2a3637e420b196a104f
targetUpstreamSha: 022a5605792996d967e05acda7091c2bd9b915db

+++ mojo/public/cpp/bindings/shared_remote.h
@@ -22,6 +22,8 @@
 #include "mojo/public/cpp/bindings/remote.h"
 #include "mojo/public/cpp/bindings/runtime_features.h"
 
+#include "base/record_replay.h"
+
 namespace mojo {
 
 namespace internal {
@@ -187,6 +189,13 @@ class SharedRemoteBase
 
     void DeleteOnCorrectThread() const {
       if (!task_runner_->RunsTasksInCurrentSequence()) {
+        // Sequenced task runners don't support posting tasks at non-deterministic
+        // points when recording/replaying, so leak the remote instead.
+        if (recordreplay::AreEventsDisallowed() &&
+            recordreplay::FeatureEnabled("leak-references", "RemoteWrapper::DeleteOnCorrectThread")) {
+          return;
+        }
+
         // NOTE: This is only called when there are no more references to
         // |this|, so binding it unretained is both safe and necessary.
         task_runner_->DeleteSoon(FROM_HERE, this);
